move sitemap logic to trait
